refactor: 메시지 처리 클래스 패키지 변경
