+++ src/main/java/com/codenear/butterfly/auth/application/AuthService.java
@@ -5,6 +5,7 @@ import com.codenear.butterfly.auth.application.email.EmailRegisterService;
 import com.codenear.butterfly.auth.application.jwt.JwtService;
 import com.codenear.butterfly.auth.domain.dto.AuthRequestDTO;
 import com.codenear.butterfly.auth.domain.dto.CustomUserDetails;
+import com.codenear.butterfly.auth.exception.message.MessageUtil;
 import com.codenear.butterfly.member.domain.Grade;
 import com.codenear.butterfly.member.domain.Member;
 import com.codenear.butterfly.member.domain.Platform;
@@ -29,10 +30,10 @@ public class AuthService {
     private final EmailRegisterService emailRegisterService;
     private final EmailLoginService emailLoginService;
     private final JwtService jwtService;
-    private final MessageService messageService;
+    private final MessageUtil messageUtil;
 
     public void handleRegistration(AuthRequestDTO requestDTO) {
-        log.info(messageService.getMessage("log.registerRequest", requestDTO.getEmail()));
+        log.info(messageUtil.getMessage("log.registerRequest", requestDTO.getEmail()));
 
         try {
             Map<Platform, Runnable> platformActions = Map.of(
@@ -43,15 +44,15 @@ public class AuthService {
             Optional.ofNullable(platformActions.get(requestDTO.getPlatform()))
                     .orElseThrow(() -> new IllegalArgumentException("제공하지 않는 플랫폼입니다."))
                     .run();
-            log.info(messageService.getMessage("log.registerSuccess", requestDTO.getEmail()));
+            log.info(messageUtil.getMessage("log.registerSuccess", requestDTO.getEmail()));
         } catch (RuntimeException e) {
-            log.error(messageService.getMessage("error.emailAlreadyInUse"));
+            log.error(messageUtil.getMessage("error.emailAlreadyInUse"));
             throw e;
         }
     }
 
     public void handleLogin(AuthRequestDTO requestDTO, HttpServletResponse response) {
-        log.info(messageService.getMessage("log.loginRequest", requestDTO.getEmail(), requestDTO.getPlatform()));
+        log.info(messageUtil.getMessage("log.loginRequest", requestDTO.getEmail(), requestDTO.getPlatform()));
 
         try {
             Map<Platform, Runnable> loginActions = Map.of(
@@ -64,12 +65,12 @@ public class AuthService {
                     .orElseThrow(() -> new IllegalArgumentException("제공하지 않는 플랫폼입니다."))
                     .run();
 
-            log.info(messageService.getMessage("log.loginSuccess", requestDTO.getEmail()));
+            log.info(messageUtil.getMessage("log.loginSuccess", requestDTO.getEmail()));
         } catch (BadCredentialsException e) {
             log.error(e.getMessage());
             throw e;
         } catch (Exception e) {
-            log.error(messageService.getMessage("error.internalServerError"));
+            log.error(messageUtil.getMessage("error.internalServerError"));
             throw e;
         }
     }
@@ -86,7 +87,7 @@ public class AuthService {
 
     private void issueJwtTokens(String email, String platform, HttpServletResponse response) {
         jwtService.processTokens(email, platform, response);
-        log.info(messageService.getMessage("log.jwtCreated", email));
+        log.info(messageUtil.getMessage("log.jwtCreated", email));
     }
 
     public void registerOrLogin(AuthRequestDTO requestDTO) {
