+++ src/main/java/com/codenear/butterfly/auth/application/email/CustomUserDetailsService.java
@@ -1,6 +1,6 @@
 package com.codenear.butterfly.auth.application.email;
 
-import com.codenear.butterfly.auth.application.MessageService;
+import com.codenear.butterfly.auth.exception.message.MessageUtil;
 import com.codenear.butterfly.auth.domain.dto.CustomUserDetails;
 import com.codenear.butterfly.member.domain.Member;
 import com.codenear.butterfly.member.domain.repository.MemberRepository;
@@ -10,20 +10,18 @@ import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
 
-import java.util.stream.Collectors;
-
 @Service
 @RequiredArgsConstructor
 public class CustomUserDetailsService implements UserDetailsService {
 
     private final MemberRepository memberRepository;
-    private final MessageService messageService;
+    private final MessageUtil messageUtil;
 
     @Override
     public UserDetails loadUserByUsername(String email) throws UsernameNotFoundException {
         Member member = memberRepository.findByEmail(email)
                 .orElseThrow(() -> new UsernameNotFoundException(
-                        messageService.getMessage("error.userNotFound", email)
+                        messageUtil.getMessage("error.userNotFound", email)
                 ));
 
         return new CustomUserDetails(
