+++ src/main/java/com/codenear/butterfly/auth/application/email/EmailLoginService.java
@@ -1,6 +1,6 @@
 package com.codenear.butterfly.auth.application.email;
 
-import com.codenear.butterfly.auth.application.MessageService;
+import com.codenear.butterfly.auth.exception.message.MessageUtil;
 import com.codenear.butterfly.auth.domain.dto.CustomUserDetails;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -14,19 +14,19 @@ import org.springframework.stereotype.Service;
 public class EmailLoginService {
 
     private final CustomUserDetailsService userDetailsService;
-    private final MessageService messageService;
+    private final MessageUtil messageUtil;
     private final PasswordEncoder passwordEncoder;
 
     public CustomUserDetails login(String email, String password) {
-        log.info(messageService.getMessage("log.loginAttempt", email));
+        log.info(messageUtil.getMessage("log.loginAttempt", email));
 
         CustomUserDetails userDetails = (CustomUserDetails) userDetailsService.loadUserByUsername(email);
 
         if (!passwordEncoder.matches(password, userDetails.getPassword())) {
-            throw new BadCredentialsException(messageService.getMessage("error.badCredentials"));
+            throw new BadCredentialsException(messageUtil.getMessage("error.badCredentials"));
         }
 
-        log.info(messageService.getMessage("log.loginSuccess", email));
+        log.info(messageUtil.getMessage("log.loginSuccess", email));
         return userDetails;
     }
 }
