+++ src/main/java/com/codenear/butterfly/auth/presentation/AuthController.java
@@ -1,7 +1,7 @@
 package com.codenear.butterfly.auth.presentation;
 
 import com.codenear.butterfly.auth.application.AuthService;
-import com.codenear.butterfly.auth.application.MessageService;
+import com.codenear.butterfly.auth.exception.message.MessageUtil;
 import com.codenear.butterfly.auth.application.jwt.JwtService;
 import com.codenear.butterfly.auth.domain.dto.AuthRequestDTO;
 import jakarta.servlet.http.HttpServletRequest;
@@ -10,12 +10,10 @@ import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.authentication.BadCredentialsException;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import org.springframework.web.server.ResponseStatusException;
 
 @RestController
 @RequiredArgsConstructor
@@ -23,19 +21,19 @@ import org.springframework.web.server.ResponseStatusException;
 public class AuthController implements AuthControllerSwagger {
     private final AuthService authService;
     private final JwtService jwtService;
-    private final MessageService messageService;
+    private final MessageUtil messageUtil;
 
     @PostMapping("/register")
     public ResponseEntity<String> register(@Valid @RequestBody AuthRequestDTO requestDTO) {
         authService.handleRegistration(requestDTO);
         return ResponseEntity.status(HttpStatus.CREATED)
-                .body(messageService.getMessage("success.register", requestDTO.getEmail()));
+                .body(messageUtil.getMessage("success.register", requestDTO.getEmail()));
     }
 
     @PostMapping("/login")
     public ResponseEntity<?> login(@Valid @RequestBody AuthRequestDTO requestDTO, HttpServletResponse response) {
         authService.handleLogin(requestDTO, response);
-        return ResponseEntity.ok(messageService.getMessage("log.loginSuccess", requestDTO.getEmail()));
+        return ResponseEntity.ok(messageUtil.getMessage("log.loginSuccess", requestDTO.getEmail()));
     }
 
     @PostMapping("/reissue")
